fix: use deep equality for property history comparisons
Replace shallow valuesEqual with recursive deepEqual to handle nested
objects (e.g. { bounds: { north: 40, south: 30 } }). Extract the
existing deepEqual from serialization.ts into a shared deep-equal.ts
module used by both.
