+++ noodles-editor/src/noodles/utils/deep-equal.ts
@@ -1,0 +1,9 @@
+export function deepEqual(a: unknown, b: unknown): boolean {
+  if (a === b) return true
+  if (a === null || b === null || typeof a !== 'object' || typeof b !== 'object') return false
+  if (Array.isArray(a) !== Array.isArray(b)) return false
+  const keysA = Object.keys(a as object)
+  const keysB = Object.keys(b as object)
+  if (keysA.length !== keysB.length) return false
+  return keysA.every(k => deepEqual((a as any)[k], (b as any)[k]))
+}
