+++ noodles-editor/src/noodles/utils/serialization.ts
@@ -100,15 +100,7 @@ export type SerializeNodesOptions = {
   forClipboard?: boolean
 }
 
-function deepEqual(a: unknown, b: unknown): boolean {
-  if (a === b) return true
-  if (a === null || b === null || typeof a !== 'object' || typeof b !== 'object') return false
-  if (Array.isArray(a) !== Array.isArray(b)) return false
-  const keysA = Object.keys(a as object)
-  const keysB = Object.keys(b as object)
-  if (keysA.length !== keysB.length) return false
-  return keysA.every(k => deepEqual((a as any)[k], (b as any)[k]))
-}
+import { deepEqual } from './deep-equal'
 
 // Check if two sets have the same elements
 function setsEqual(a: Set<string>, b: Set<string>): boolean {
